github_ready/requirements.txt
flask==3.0.0
flask-cors==5.0.0
tensorflow-cpu==2.12.0
opencv-python-headless==4.8.0.74
numpy==1.23.5
gunicorn==21.2.0
